+++ app/models/claim.rb
@@ -24,6 +24,7 @@ class Claim < ApplicationRecord
     :student_loan_country,
     :student_loan_courses,
     :student_loan_start_date,
+    :has_masters_doctoral_loan,
     :postgraduate_masters_loan,
     :postgraduate_doctoral_loan,
     :email_address,
@@ -57,6 +58,7 @@ class Claim < ApplicationRecord
     student_loan_country: false,
     student_loan_courses: false,
     student_loan_start_date: false,
+    has_masters_doctoral_loan: false,
     postgraduate_masters_loan: false,
     postgraduate_doctoral_loan: false,
     email_address: true,
@@ -93,9 +95,10 @@ class Claim < ApplicationRecord
   DECISION_DEADLINE = 12.weeks
   DECISION_DEADLINE_WARNING_POINT = 2.weeks
   ATTRIBUTE_DEPENDENCIES = {
-    "has_student_loan" => ["student_loan_country", "postgraduate_masters_loan", "postgraduate_doctoral_loan"],
+    "has_student_loan" => ["student_loan_country"],
     "student_loan_country" => ["student_loan_courses"],
     "student_loan_courses" => ["student_loan_start_date"],
+    "has_masters_doctoral_loan" => ["postgraduate_masters_loan", "postgraduate_doctoral_loan"],
     "bank_or_building_society" => ["banking_name", "bank_account_number", "bank_sort_code", "building_society_roll_number"],
     "provide_mobile_number" => ["mobile_number"]
   }.freeze
@@ -185,8 +188,10 @@ class Claim < ApplicationRecord
   validates :student_loan_start_date, on: [:"student-loan-start-date"], presence: {message: ->(object, data) { I18n.t("validation_errors.student_loan_start_date.#{object.student_loan_courses}") }}
   validates :student_loan_plan, on: [:submit], presence: {message: "We have not been able determined your student loan repayment plan. Answer all questions about your student loan."}
   validates :student_loan_plan, on: [:amendment], inclusion: {in: [Claim::NO_STUDENT_LOAN], message: "You can’t amend the student loan plan type because the claimant said they are no longer paying off their student loan"}, if: :no_student_loan?
-  validates :postgraduate_masters_loan, on: [:"masters-loan", :submit], inclusion: {in: [true, false], message: "Select yes if you have a Postgraduate Master Loan taken out on or after 1st August 2016"}, if: -> { has_student_loan? }
-  validates :postgraduate_doctoral_loan, on: [:"doctoral-loan", :submit], inclusion: {in: [true, false], message: "Select yes if you have a Postgraduate Doctoral Loan taken out on or after 1st August 2018"}, if: -> { has_student_loan? }
+
+  validates :has_masters_doctoral_loan, on: [:"masters-doctoral-loan", :submit], inclusion: {in: [true, false], message: "Select yes if you have a postgraduate masters and/or doctoral loan"}
+  validates :postgraduate_masters_loan, on: [:"masters-loan", :submit], inclusion: {in: [true, false], message: "Select yes if you have a Postgraduate Master Loan taken out on or after 1st August 2016"}, if: -> { has_masters_doctoral_loan? }
+  validates :postgraduate_doctoral_loan, on: [:"doctoral-loan", :submit], inclusion: {in: [true, false], message: "Select yes if you have a Postgraduate Doctoral Loan taken out on or after 1st August 2018"}, if: -> { has_masters_doctoral_loan? }
 
   validates :email_address, on: [:"email-address", :submit], presence: {message: "Enter an email address"}
   validates :email_address, format: {with: URI::MailTo::EMAIL_REGEXP, message: "Enter an email in the format name@example.com"},
@@ -286,6 +291,10 @@ class Claim < ApplicationRecord
 
   def no_student_loan?
     !has_student_loan?
+  end
+
+  def no_masters_doctoral_loan?
+    !has_masters_doctoral_loan?
   end
 
   def student_loan_country_with_one_plan?