+++ app/models/student_loans/slug_sequence.rb
@@ -32,6 +32,7 @@ module StudentLoans
       "student-loan-how-many-courses",
       "student-loan-start-date",
       "student-loan-amount",
+      "masters-doctoral-loan",
       "masters-loan",
       "doctoral-loan",
       "email-address",
@@ -58,8 +59,8 @@ module StudentLoans
         sequence.delete("student-loan-country") if claim.no_student_loan?
         sequence.delete("student-loan-how-many-courses") if claim.no_student_loan? || claim.student_loan_country_with_one_plan?
         sequence.delete("student-loan-start-date") if claim.no_student_loan? || claim.student_loan_country_with_one_plan?
-        sequence.delete("masters-loan") if claim.no_student_loan?
-        sequence.delete("doctoral-loan") if claim.no_student_loan?
+        sequence.delete("masters-loan") if claim.no_masters_doctoral_loan?
+        sequence.delete("doctoral-loan") if claim.no_masters_doctoral_loan?
         sequence.delete("personal-bank-account") if claim.bank_or_building_society == "building_society"
         sequence.delete("building-society-account") if claim.bank_or_building_society == "personal_bank_account"
         sequence.delete("mobile-number") if claim.provide_mobile_number == false