+++ db/migrate/20210826080653_add_has_masters_doctoral_loan_to_claim.rb
@@ -1,0 +1,5 @@
+class AddHasMastersDoctoralLoanToClaim < ActiveRecord::Migration[6.0]
+  def change
+    add_column :claims, :has_masters_doctoral_loan, :boolean
+  end
+end