+++ spec/features/maths_and_physics_claim_spec.rb
@@ -116,6 +116,14 @@ RSpec.feature "Maths & Physics claims" do
       expect(claim.student_loan_start_date).to eq(StudentLoan::BEFORE_1_SEPT_2012)
       expect(claim.student_loan_plan).to eq(StudentLoan::PLAN_1)
 
+      # - Are you currently paying off your masters/doctoral loan
+      expect(page).to have_text(I18n.t("questions.has_masters_and_or_doctoral_loan"))
+
+      choose "Yes"
+      click_on "Continue"
+
+      expect(claim.reload.has_masters_doctoral_loan).to eql true
+
       # - Did you take out a postgraduate masters loan on or after 1 August 2016
       expect(page).to have_text(I18n.t("questions.postgraduate_masters_loan"))