+++ spec/models/maths_and_physics/slug_sequence_spec.rb
@@ -57,15 +57,11 @@ RSpec.describe MathsAndPhysics::SlugSequence do
       expect(slug_sequence.slugs).not_to include("student-loan-country")
       expect(slug_sequence.slugs).not_to include("student-loan-how-many-courses")
       expect(slug_sequence.slugs).not_to include("student-loan-start-date")
-      expect(slug_sequence.slugs).not_to include("masters-loan")
-      expect(slug_sequence.slugs).not_to include("doctoral-loan")
 
       claim.has_student_loan = true
       expect(slug_sequence.slugs).to include("student-loan-country")
       expect(slug_sequence.slugs).to include("student-loan-how-many-courses")
       expect(slug_sequence.slugs).to include("student-loan-start-date")
-      expect(slug_sequence.slugs).to include("masters-loan")
-      expect(slug_sequence.slugs).to include("doctoral-loan")
     end
 
     it "excludes the remaining student loan plan slugs if the claimant received their student loan in a Plan 1 country" do
@@ -84,6 +80,16 @@ RSpec.describe MathsAndPhysics::SlugSequence do
         expect(slug_sequence.slugs).to include("student-loan-how-many-courses")
         expect(slug_sequence.slugs).to include("student-loan-start-date")
       end
+    end
+
+    it "excludes postgraduate masters and doctoral loan plan slugs if the claimant is not paying off a postgraduate masters and/or doctoral loan" do
+      claim.has_masters_doctoral_loan = false
+      expect(slug_sequence.slugs).not_to include("masters-loan")
+      expect(slug_sequence.slugs).not_to include("doctoral-loan")
+
+      claim.has_masters_doctoral_loan = true
+      expect(slug_sequence.slugs).to include("masters-loan")
+      expect(slug_sequence.slugs).to include("doctoral-loan")
     end
 
     context "when claim payment details are 'personal bank account'" do